Created component for favorite button.
