+++ src/pages/offer-page/offer-page.tsx
@@ -1,17 +1,15 @@
 import { Helmet } from 'react-helmet-async';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import classNames from 'classnames';
 import { useEffect } from 'react';
 import { capitalizeFirstLetter, getRating } from '@utils/common';
 import { useAppDispatch, useAppSelector } from '@hooks/index';
-import { AppRoute, AuthorizationStatus, IMAGES_COUNT, NEAR_OFFERS_COUNT, RequestStatus } from '@const';
+import { AuthorizationStatus, IMAGES_COUNT, NEAR_OFFERS_COUNT, RequestStatus } from '@const';
 import { authSelectors } from '@store/slices/auth';
 import { fetchCommentsAction } from '@store/thunks/comments';
 import { offerSelectors } from '@store/slices/offer';
 import { fetchNearByOffersAction, fetchOfferByIdAction } from '@store/thunks/offers';
 import { commentsSelectors } from '@store/slices/comments';
-import { changeFavoriteAction } from '@store/thunks/favorites';
-import { getToken } from '@services/token';
 import Container from '@components/container/container';
 import OfferCommentForm from '@components/offer-comment-form/offer-comment-form';
 import ListComments from '@components/list-comments/list-comments';
@@ -19,17 +17,16 @@ import Map from '@components/map/map';
 import ListOffers from '@components/list-offers/list-offers';
 import NotFoundPage from '@pages/not-found-page/not-found-page';
 import Loader from '@components/loader/loader';
+import FavoriteButton from '@components/favorite-button/favorite-button';
 
 function OfferPage(): JSX.Element | undefined {
   const { offerId } = useParams();
   const dispatch = useAppDispatch();
-  const navigate = useNavigate();
   const authorizationStatus = useAppSelector(authSelectors.selectAuthorizationStatus);
   const status = useAppSelector(offerSelectors.selectOfferStatus);
   const offerInfo = useAppSelector(offerSelectors.selectOffer);
   const nearOffers = useAppSelector(offerSelectors.selectNearByOffers);
   const comments = useAppSelector(commentsSelectors.selectComments);
-  const token = getToken();
 
   useEffect(() => {
     dispatch(fetchOfferByIdAction(offerId as string));
@@ -48,14 +45,6 @@ function OfferPage(): JSX.Element | undefined {
       <NotFoundPage />
     );
   }
-
-  const handleFavoriteChange = () => {
-    if (!token) {
-      return navigate(AppRoute.Login);
-    }
-
-    dispatch(changeFavoriteAction({offerId: offerInfo?.id, status: Number(!offerInfo.isFavorite)}));
-  };
 
   const threeNearOffers = nearOffers.slice(0, NEAR_OFFERS_COUNT);
   const nearOffersAndCurrent = [offerInfo, ...threeNearOffers];
@@ -84,16 +73,7 @@ function OfferPage(): JSX.Element | undefined {
 
             <div className="offer__name-wrapper">
               <h1 className="offer__name">{offerInfo?.title}</h1>
-              <button
-                className={classNames('offer__bookmark-button', 'button', {'offer__bookmark-button--active': offerInfo?.isFavorite})}
-                type="button"
-                onClick={handleFavoriteChange}
-              >
-                <svg className="offer__bookmark-icon" width={31} height={33}>
-                  <use xlinkHref="#icon-bookmark" />
-                </svg>
-                <span className="visually-hidden">To bookmarks</span>
-              </button>
+              <FavoriteButton extraClass='offer' offerId={offerInfo.id} isFavorite={offerInfo.isFavorite} width={31} height={33} />
             </div>
             <div className="offer__rating rating">
               <div className="offer__stars rating__stars">
